sig: move sig config to go/pkg/sig
