+++ go/pkg/sig/sigconfig/conf.go
