+++ go/pkg/sig/sigconfig/conf_test.go
