+++ go/pkg/sig/sigconfig/sample.go
