+++ go/sig/BUILD.bazel
@@ -16,12 +16,12 @@ go_library(
         "//go/lib/serrors:go_default_library",
         "//go/lib/sigdisp:go_default_library",
         "//go/lib/sigjson:go_default_library",
+        "//go/pkg/sig/sigconfig:go_default_library",
         "//go/sig/egress:go_default_library",
         "//go/sig/internal/base:go_default_library",
         "//go/sig/internal/ingress:go_default_library",
         "//go/sig/internal/metrics:go_default_library",
         "//go/sig/internal/sigcmn:go_default_library",
-        "//go/sig/internal/sigconfig:go_default_library",
         "//go/sig/internal/xnet:go_default_library",
         "@com_github_pelletier_go_toml//:go_default_library",
         "@com_github_syndtr_gocapability//capability:go_default_library",
