+++ go/sig/internal/sigcmn/common.go
@@ -31,7 +31,7 @@ import (
 	"github.com/scionproto/scion/go/lib/serrors"
 	"github.com/scionproto/scion/go/lib/snet"
 	"github.com/scionproto/scion/go/lib/sock/reliable"
-	"github.com/scionproto/scion/go/sig/internal/sigconfig"
+	"github.com/scionproto/scion/go/pkg/sig/sigconfig"
 	"github.com/scionproto/scion/go/sig/internal/snetmigrate"
 )
